Extract Methods gem listener

# Conflicts:
#	lib/tapioca/compilers/symbol_table_compiler.rb
#	lib/tapioca/internal.rb
