+++ lib/tapioca/internal.rb
@@ -19,6 +19,7 @@ require "tapioca/symbol_loader"
 require "tapioca/compilers/gem/events"
 require "tapioca/compilers/gem/listeners/base"
 require "tapioca/compilers/gem/listeners/dynamic_mixins"
+require "tapioca/compilers/gem/listeners/methods"
 require "tapioca/compilers/gem/listeners/mixins"
 require "tapioca/compilers/gem/listeners/sorbet_enums"
 require "tapioca/compilers/gem/listeners/sorbet_helpers"
